Bias the MEMC moves

+++ src/ConfigSetup.h
@@ -243,7 +243,7 @@ struct MEMCVal {
 
 struct CFCMCVal {
   bool enable, readWindow, readRelaxSteps, readHistFlatness;
-  long window, relaxSteps;
+  int window, relaxSteps;
   double histFlatness;
   CFCMCVal(void) {
     readWindow = readRelaxSteps = readHistFlatness = false;
